Drop required_action_json column; derive the metric from tool_executions
The aggregate is one GROUP BY away from existing data
(tool_executions.agent_execution_id + tool_name + success), and a
denormalized field with no in-product reader is a second source of truth
waiting to disagree with the first. Removes the column, migration, and
finish-time computation; keeps the fix-1 schema-guidance tests under
their own file.
